OTP Encryption
The program uses a basic one-time pad encryption method to easily encrypt and decrypt a given file. The user provides a key file and the phrase they wish to encrypt, The program will output the encrypted text to a file and provide a description key. These two files are given to the decryption program which will return the plain text.